Добавил модульность
